fix: round and cast due days/hours/minutes to integers to prevent floating-point display

+++ app/Services/GoogleClassroomService.php
@@ -243,7 +243,7 @@ class GoogleClassroomService
 
     protected function calculatePriority(Carbon $dueDateTime): string
     {
-        $daysUntil = now()->diffInDays($dueDateTime, false);
+        $daysUntil = (int) now()->diffInDays($dueDateTime, false);
 
         if ($daysUntil < 0) {
             return 'tinggi';
